requirements.txt
matplotlib>=3.7.0
pandas>=2.0.0
PyYAML>=6.0
